chore(btc_server): rename cordinator.rs -> coordinator.rs
