+++ bin/btc-server/src/coordinator.rs
